requirements.txt
fastapi
sqlalchemy 
pymysql
uvicorn
pydantic
python-dotenv
mysql-connector-python